split switcher plugins
