+++ system/autoware_command_mode_switcher/CMakeLists.txt
@@ -3,7 +3,6 @@ project(autoware_command_mode_switcher)
 
 find_package(autoware_cmake REQUIRED)
 autoware_package()
-pluginlib_export_plugin_description_file(${PROJECT_NAME} plugins.xml)
 
 ament_auto_add_library(${PROJECT_NAME}_common SHARED
   "src/common/command_container.cpp"
@@ -13,17 +12,7 @@ ament_auto_add_library(${PROJECT_NAME}_common SHARED
   "src/common/selector_interface.cpp"
   "src/command_mode_switcher.cpp"
 )
-ament_auto_add_library(${PROJECT_NAME}_plugin SHARED
-  "src/plugins/stop.cpp"
-  "src/plugins/autonomous.cpp"
-  "src/plugins/local.cpp"
-  "src/plugins/remote.cpp"
-  "src/plugins/emergency_stop.cpp"
-  "src/plugins/comfortable_stop.cpp"
-  "src/plugins/pull_over.cpp"
-)
 target_include_directories(${PROJECT_NAME}_common PRIVATE "src")
-target_include_directories(${PROJECT_NAME}_plugin PRIVATE "src")
 
 rclcpp_components_register_node(${PROJECT_NAME}_common
   PLUGIN "autoware::command_mode_switcher::CommandModeSwitcher"
