+++ system/autoware_command_mode_switcher/src/common/command_container.hpp
@@ -15,7 +15,7 @@
 #ifndef COMMON__COMMAND_CONTAINER_HPP_
 #define COMMON__COMMAND_CONTAINER_HPP_
 
-#include "common/command_plugin.hpp"
+#include "autoware_command_mode_switcher/command_plugin.hpp"
 #include "common/command_status.hpp"
 
 #include <memory>
