+++ system/autoware_command_mode_switcher/src/common/manual.hpp
@@ -15,7 +15,7 @@
 #ifndef COMMON__MANUAL_HPP_
 #define COMMON__MANUAL_HPP_
 
-#include "common/command_plugin.hpp"
+#include "autoware_command_mode_switcher/command_plugin.hpp"
 
 #include <autoware_command_mode_types/constants/modes.hpp>
 #include <autoware_command_mode_types/constants/sources.hpp>
