+++ system/autoware_command_mode_switcher_plugins/plugins.xml
@@ -1,4 +1,4 @@
-<library path="autoware_command_mode_switcher_plugin">
+<library path="autoware_command_mode_switcher_plugins">
   <class base_class_type="autoware::command_mode_switcher::CommandPlugin" type="autoware::command_mode_switcher::StopSwitcher"/>
   <class base_class_type="autoware::command_mode_switcher::CommandPlugin" type="autoware::command_mode_switcher::AutonomousSwitcher"/>
   <class base_class_type="autoware::command_mode_switcher::CommandPlugin" type="autoware::command_mode_switcher::LocalSwitcher"/>
