+++ system/autoware_command_mode_switcher_plugins/src/autonomous.cpp
