+++ system/autoware_command_mode_switcher_plugins/src/comfortable_stop.cpp
@@ -12,29 +12,18 @@
 //  See the License for the specific language governing permissions and
 //  limitations under the License.
 
-#ifndef PLUGINS__REMOTE_HPP_
-#define PLUGINS__REMOTE_HPP_
-
-#include "common/command_plugin.hpp"
-
-#include <autoware_command_mode_types/constants/modes.hpp>
-#include <autoware_command_mode_types/constants/sources.hpp>
+#include "comfortable_stop.hpp"
 
 namespace autoware::command_mode_switcher
 {
 
-class RemoteSwitcher : public ControlCommandPlugin
+void ComfortableStopSwitcher::initialize()
 {
-public:
-  uint16_t mode() const override { return autoware::command_mode_types::modes::remote; }
-  uint16_t source() const override { return autoware::command_mode_types::sources::remote; }
-  bool autoware_control() const override { return true; }
-  void initialize() override;
-
-  bool get_transition_available() override { return true; }
-  bool get_transition_completed() override { return true; }
-};
+}
 
 }  // namespace autoware::command_mode_switcher
 
-#endif  // PLUGINS__REMOTE_HPP_
+#include <pluginlib/class_list_macros.hpp>
+PLUGINLIB_EXPORT_CLASS(
+  autoware::command_mode_switcher::ComfortableStopSwitcher,
+  autoware::command_mode_switcher::CommandPlugin)
