+++ system/autoware_command_mode_switcher_plugins/src/pull_over.hpp
@@ -12,23 +12,25 @@
 //  See the License for the specific language governing permissions and
 //  limitations under the License.
 
-#include "local.hpp"
+#ifndef PULL_OVER_HPP_
+#define PULL_OVER_HPP_
+
+#include <autoware_command_mode_switcher/command_plugin.hpp>
+#include <autoware_command_mode_types/constants/modes.hpp>
+#include <autoware_command_mode_types/constants/sources.hpp>
 
 namespace autoware::command_mode_switcher
 {
 
-void LocalSwitcher::initialize()
+class PullOverSwitcher : public ControlCommandPlugin
 {
-  /*
-  set_mode_continuable(false);
-  set_mode_available(false);
-  set_ctrl_available(true);
-  set_transition_completed(true);
-  */
-}
+public:
+  uint16_t mode() const override { return autoware::command_mode_types::modes::pull_over; }
+  uint16_t source() const override { return autoware::command_mode_types::sources::main; }
+  bool autoware_control() const override { return true; }
+  void initialize() override;
+};
 
 }  // namespace autoware::command_mode_switcher
 
-#include <pluginlib/class_list_macros.hpp>
-PLUGINLIB_EXPORT_CLASS(
-  autoware::command_mode_switcher::LocalSwitcher, autoware::command_mode_switcher::CommandPlugin)
+#endif  // PULL_OVER_HPP_
