+++ system/autoware_command_mode_switcher_plugins/src/remote.cpp
@@ -12,29 +12,17 @@
 //  See the License for the specific language governing permissions and
 //  limitations under the License.
 
-#ifndef PLUGINS__STOP_HPP_
-#define PLUGINS__STOP_HPP_
-
-#include "common/command_plugin.hpp"
-
-#include <autoware_command_mode_types/constants/modes.hpp>
-#include <autoware_command_mode_types/constants/sources.hpp>
+#include "remote.hpp"
 
 namespace autoware::command_mode_switcher
 {
 
-class StopSwitcher : public ControlCommandPlugin
+void RemoteSwitcher::initialize()
 {
-public:
-  uint16_t mode() const override { return autoware::command_mode_types::modes::stop; }
-  uint16_t source() const override { return autoware::command_mode_types::sources::stop; }
-  bool autoware_control() const override { return true; }
-  void initialize() override;
-
-  bool get_transition_available() override { return true; }
-  bool get_transition_completed() override { return true; }
-};
+}
 
 }  // namespace autoware::command_mode_switcher
 
-#endif  // PLUGINS__STOP_HPP_
+#include <pluginlib/class_list_macros.hpp>
+PLUGINLIB_EXPORT_CLASS(
+  autoware::command_mode_switcher::RemoteSwitcher, autoware::command_mode_switcher::CommandPlugin)
