+++ system/autoware_command_mode_switcher_plugins/src/stop.hpp
@@ -12,24 +12,28 @@
 //  See the License for the specific language governing permissions and
 //  limitations under the License.
 
-#include "comfortable_stop.hpp"
+#ifndef STOP_HPP_
+#define STOP_HPP_
+
+#include <autoware_command_mode_switcher/command_plugin.hpp>
+#include <autoware_command_mode_types/constants/modes.hpp>
+#include <autoware_command_mode_types/constants/sources.hpp>
 
 namespace autoware::command_mode_switcher
 {
 
-void ComfortableStopSwitcher::initialize()
+class StopSwitcher : public ControlCommandPlugin
 {
-  /*
-  set_mode_continuable(false);
-  set_mode_available(false);
-  set_ctrl_available(true);
-  set_transition_completed(true);
-  */
-}
+public:
+  uint16_t mode() const override { return autoware::command_mode_types::modes::stop; }
+  uint16_t source() const override { return autoware::command_mode_types::sources::stop; }
+  bool autoware_control() const override { return true; }
+  void initialize() override;
+
+  bool get_transition_available() override { return true; }
+  bool get_transition_completed() override { return true; }
+};
 
 }  // namespace autoware::command_mode_switcher
 
-#include <pluginlib/class_list_macros.hpp>
-PLUGINLIB_EXPORT_CLASS(
-  autoware::command_mode_switcher::ComfortableStopSwitcher,
-  autoware::command_mode_switcher::CommandPlugin)
+#endif  // STOP_HPP_
